feat: implement async timer replacing the old threaded model
Converted the timer from running on a different thread to being async and run on the same thread.
The timer now also takes use of g_timeout_add to add an interrupt/schedule a call to update the timer after a second.
Progress circle now smoothly animates and updates with the frame rate of the display ( every time the window renders a frame ).
Added completion notifications on session completion.